requirements.txt
langchain
faiss-cpu
streamlit
python-dotenv
langchain-community
PyMuPDF
transformers
torch